+++ actioncable/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Add notification hooks for connection events
+
+    `connect.action_cable` and `disconnect.action_cable` notifications are added.
+
+    *Sergey Ponomarev*
 
 
 Please check [6-0-stable](https://github.com/rails/rails/blob/6-0-stable/actioncable/CHANGELOG.md) for previous changes.